source/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "syco"
version = "0.1.0"
description = "Epistemic sycophancy benchmark for LLMs"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "openai>=1.50",
    "python-dotenv>=1.0",
    "tenacity>=8.3",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "matplotlib>=3.9",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "pydantic>=2.7",
    "tabulate>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
syco = "syco.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
